refactor: Add tags, share code and update readme
