+++ src/DataLoader.DynamicMvvm/DataLoader.DynamicMvvm.csproj
@@ -10,7 +10,7 @@
 		<Description>Recipes for loading asynchronous data in a MVVM context.</Description>
 		<GenerateDocumentationFile>true</GenerateDocumentationFile>
 		<PackageReadmeFile>README.md</PackageReadmeFile>
-		<PackageTags>dataloader;async;mvvm;xamarin;ios;android;uwp;chinook</PackageTags>
+		<PackageTags>dataloader;async;mvvm;xamarin;ios;android;uwp;chinook;maui;winui</PackageTags>
 		<PackageLicenseExpression>Apache-2.0</PackageLicenseExpression>
 		<PackageProjectUrl>https://github.com/nventive/Chinook.DataLoader</PackageProjectUrl>
