+++ src/Shared/Shared.projitems
@@ -9,7 +9,9 @@
     <Import_RootNamespace>Shared</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)DataLoaderView.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)DataLoaderView.Properties.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)DataLoaderViewControllerBase.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)DataLoaderViewState.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)IDataLoaderViewDelegate.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)PreserveAttribute.cs" />
